requirements.txt
Flask==3.0.0
Flask-Mail==0.9.1
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
gunicorn==21.2.0
python-dotenv==1.0.0
Pillow==10.1.0
email-validator==2.1.0
pymongo==4.6.0
dnspython>=2.4.0
setuptools>=65.0.0
wheel>=0.37.0
